Move: api test folder to app root folder
