+++ test/api/departments/put.test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const server = require('../../../../server.js');
-const Department = require('../../../department.model');
+const server = require('../../../server');
+const Department = require('../../../models/department.model');
 
 chai.use(chaiHttp);
